Rename key_test file and update the expected key
